add utility function to copy and iteration over code elements
